+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.7.0] - Unreleased
+
+### Added
+
+- Spell rotation combo box in Spell Queue for per-character setting
+- Spell rotation character setting is preserved in saved state
+- New option for `Skip Spells on Cooldown` for `Spell Macros` (default is `Enabled`)
+- More accessibility key for checkboxes in `Spell Macro` settings
+
+### Changed
+
+- `Spell Rotation Mode` renamed `Default Spell Queue Rotation` to better describe it can be overriden
+- `UserSettings` are now version `1.5`
+
+### Fixed
+
+- Some staff line changes
+
 ## [4.6.1] - 2023-06-03
 
 ### Added